+++ tests/BootstrapAggregatorTest.php
@@ -3,7 +3,6 @@
 namespace Rubix\ML\Tests;
 
 use Rubix\ML\Learner;
-use Rubix\ML\Parallel;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Datasets\Unlabeled;
@@ -44,7 +43,6 @@ class BootstrapAggregatorTest extends TestCase
     {
         $this->assertInstanceOf(BootstrapAggregator::class, $this->estimator);
         $this->assertInstanceOf(Learner::class, $this->estimator);
-        $this->assertInstanceOf(Parallel::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);